delete confirmation

+++ app/javascript/controllers/tabular_controller.js
@@ -1,4 +1,3 @@
-import { Controller } from "stimulus"
 import ApplicationController from "./application_controller"
 import { headers } from '../src/js/csrf'
 
@@ -6,20 +5,19 @@ export default class extends ApplicationController {
 
   // action
   destroy(event) {
-    const user = this.data.get('id')
-
-    let attachmentId = event.currentTarget.dataset.uploadId
-    let uploadHolder = event.currentTarget.closest('.attachment__holder')
-    fetch('/users/' + user, {
-      method: 'delete',
-      headers: headers,
-    }).then(response => {
-      if (response.ok) {
-        this.element.remove()
-      } else {
-        console.error(response)
-      }
-    }).catch(err => console.error('we got an error: ', err))
+    if (confirm('Are you sure?')) {
+      const user = this.data.get('id')
+      fetch('/users/' + user, {
+        method: 'delete',
+        headers: headers,
+      }).then(response => {
+        if (response.ok) {
+          this.element.remove()
+        } else {
+          console.error(response)
+        }
+      }).catch(err => console.error('we got an error: ', err))
+    }
   }
 
 }
